FEAT: Implement a careful “MaybeUninit” using ManuallyDrop
This is a cautious version of MaybeUninit, since we don't have one in
libstd, based on ManuallyDrop.

This moves ArrayVec to a nice, no size overhead implementation by
default. We use Rust version sniffing (build script) to automatically
use this for new enough Rust versions.

This doesn't kill nodrop unfortunately, it still remains as the
fallback.