FIX: Use raw pointer casts for MaybeUninit
The nodrop/fallback code is as it was, it uses the Array trait accessors
as_ptr/as_ptr_mut; but we really want to do this properly with a raw
pointer cast according to discussion with arielb.

The idea is that we don't take a reference to a partially uninitialized
array value -- we just extract a raw pointer to the first array element
instead.