Fixed miri errors for `fn drain(...)`.
This required to rewrite `struct Drain` because old implementation had aliasing pointers in it which violated [Stacked Borrows][1].
New implementation separately borrows len field and parts of the slice to prevent aliasing.

Also, we don't use slice iteration anymore because stacked borrows conflict with `ptr::copy` which we use to move the tail. We can wrote it valid for stacked borrows by using 2 memmoves (1 from tail beginning to removed slice and 1 from rest of the tail to beginning of the tail) but it is more complex to read and understand, more error prone and less effecient.

This commit causes divergence with Rust `alloc::vec::Vec` implementation but we cannot implement it same way because it is only possible if `Vec` struct stored in different allocation than its buffer which isn't possible for `ArrayVec`.

Also, added test that check drain behaviour when dropping.

[1]: https://plv.mpi-sws.org/rustbelt/stacked-borrows/