Avoid `unsafe` in `ArrayString`'s `new` and `zero_filled` methods.
Before the changes in `fn new`, the type of `MaybeUninit::uninit()`
would be inferred to be `MaybeUninit<[MaybeUninit<...>; CAP]>` which
seems somewhat unexpected.  This commit rewrites this expression so that
no `unsafe` is needed.

Similarily, before the changes in `fn zero_filled`, the type of
`MaybeUninit::zeroed()` would be inferred to be
`MaybeUninit<[MaybeUninit<...>; CAP]>`.  This commit also rewrites this
expression so that no `unsafe` is needed.